pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfg-semantic-grounding"
version = "0.1.0"
description = "Clean SWE-Bench attack/defense group-testing harness"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "rich>=13.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "joblib>=1.3",
]

[project.optional-dependencies]
pydantic = ["pydantic>=2.0"]
llm = [
  "openai>=1.0",
  "anthropic>=0.34",
  "google-generativeai>=0.7.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
